pyproject.toml
[tool.poetry]
name = "gcr-cli"
version = "0.1.0"
description = "A helper for working with GitHub classroom."
authors = ["James Turk <dev@jamesturk.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
gcr = 'gcr_cli.gcr:app'

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.6.1"
pygithub = "^1.56"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.10.0"
flake8-pyproject = "^1.1.0.post0"

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
